feat: add encoder utilities for pushdown (#2388)
This adds a new field encoder (ZoneMapFieldEncoder) that calculates
pushdown statistics and places them in the metadata.

It also changes the encoder so that it the choice of encoding is
configurable. This makes it possible for extensions to register custom
encodings. The zone maps encoder is an example of this as it is placed
in a special crate for "encodings that rely on datafusion".

It also adds some utilities for converting an `EncodedBatch` to `Bytes`
according to the lance file format. This makes it possible to go from
`RecordBatch` to `Bytes` using the lance file format.

There is not much testing for the zone maps encoder. More will come when
we add support for reading zone maps but I want to keep this PR simple
for now.