fix: fix incorrect validation logic in updater (#2408)
The updater has a check in the deletion restorer to make sure each batch
has the same size as the batch size (if using v1 files). This is not
true for the final batch because it might be smaller than the batch
size. This PR adds a unit test to regress this case and fixes the
validation logic.