fix: fix file writer which was not writing page buffers in the correct order (#2413)
Pages with multiple buffers (today just primitive pages where some of
the values are null) were being mangled as a result. This was not caught
during testing since we only had null testing in encoding unit tests
which don't go through lance files. I've added a basic test for nulls to
the python tests to catch this in the future.