refactor: refactor logical decoders into "field decoders" (#2407)
Previously we had "logical page scheduler" and "physical page
scheduler". There was no concept of a "column scheduler" or "field
scheduler" like we have on the encode path. This was going to be a
problem for pushdown and dictionary encodings. Both of these operate on
a field-level (technically dictionary encoding might someday operate on
both a page and field level but we can ignore that for now).

This PR converts the "logical page scheduler" concept into a "field
scheduler" which spans entire columns.

The result is that the struct scheduler got simpler (some of its
complexity moved to the primitive scheduler) and the list scheduler
stayed more or less as complex as it has always been (but the work got
split up a bit better). Performance-wise there should be no real
difference but this opens up the floor for pushdown, dictionary, rle,
etc.

There is also the small perk of aligning better with the encoder traits
which have always had "field encoder"