feat: add pushdown to the read path (#2444)
This PR adds quite a few pre-requisites as well:

 * We add the concept of a "filter" to the read path
* We make the selection of decoder extensible with the
`FieldDecoderStrategy` and `DecoderMiddlewareChain`
* Various changes to the decode path so that we can handle the case
where the number of rows scheduled is less than the number of rows
originally asked for
* We add Lance deserialization to `EncodedBatch` (serialization was
added in an earlier PR)
* We fix some bugs in the serialization and deserialization of
`EncodedBatch`

The basic flow is this:
* On write, we store statistics for each leaf column. These are a record
batch of min/max/null_count with a row for every X rows (X is
configurable and we should experiment with it).
* This batch is encoded into a single buffer (as a mini lance file) and
written at the end of the data section of the file. Each leaf column's
encoding is wrapped with a header that points to this buffer
* When creating decoders, every time we create a leaf column, we unwrap
the stats location and collect these.
* When creating the root decoder, after the decoder is created, we wrap
the root decoder with a filtering decoder that is aware of all of these
zone maps. At this time we go out and collect all the zone maps from the
file and decode them back into record batches. We now have a map from
field to zone map
* At schedule time, this outer decoder uses the user provided filter and
the zone maps to refine the range that the user asked for.

This is not the final PR for pushdown, there are still a few things that
need to be done:

* We need to use `LanceDfFieldDecoderStrategy` in fragment and the
python bindings
 * The way we are mapping fields to DF columns is not quite right
* Ideally we can move the zone map "initialization" to the beginning of
the scheduling process (instead of opening the file) so that we can only
inflate statistics for columns involved in the filter
 * We need to cache the inflated statistics in the cached metadata
* Ideally we can add a "context" to the decoder middleware chain so that
`LanceDfFieldDecoderStrategy` doesn't have to use a mutex

Right now pushdown is implemented as a "best effort" filter so a
`FilterExec` is still needed in any plan doing pushdown. In the future
we might want to move the `FilterExec` into the decoder. This would
close https://github.com/lancedb/lance/issues/2398