feat: scan with stable row id (#2441)
Modifies the scanner to now support two options: `with_row_id` and
`with_row_address`. Row addresses are what we used to call `_rowid`.
They are returned as the `_rowaddr` column. For backwards compatibility,
the `_rowid` column will be a copy of `_rowaddr` for datasets that
aren't using stable row ids. For those that have activated the stable
row id feature, they will have the row ids in the `_rowid` column.