feat: pickle the manifest when pickling a dataset (#2459)
This allows us to send the pickled dataset across IPC and not require a
reload of the manifest on the remote. It should be safe since manifests
are read only. It uses the existing protobuf serialization for
manifests.

If the manifest contains a dictionary field then we still need to load
the dictionary information on the remote. Perhaps we can pickle the
dictionary at some point (or, with lv2, maybe dictionary won't need to
be in manifest)