feat: allow users to create a file writer without a schema (#2558)
This is needed for some of the index refactoring work I'm doing. It can
be difficult to know the schema in advance and, in many (most) cases,
the desired file schema is the same as the data schema and so there is
no real benefit to providing it in advance.

This PR also removes the "path" argument from the rust constructor. The
path is already part of the ObjectWriter and there is no need for the
file writer to keep a second copy.