fix: add support for list fields with names other than 'item' (#2580)
In the future we can maybe add support for a few more things as well:

* User's should be able to read back into whatever schema they want
(this is present in the rust already but missing from python because
projection is missing from python, so this is mostly just tests)
* Perhaps add some kind of read option to "normalize" a schema so we
always read back the field as "item". Right now, if no schema is
provided at read time, we mirror exactly the write time schema, this
will cause non-item field names to propagate which is maybe not the best
choice. This could also be a write time option to normalize the schema
on write.