feat: add label list index (#2581)
This PR adds a new kind of scalar index, the LABEL_LIST scalar index. A
label list index can only be created on list columns (the item data type
has to be one of the data types that can be used for btree/bitmap
indices). The label list index can speed up array_has_any and
array_has_all queries.

This is useful for tag-like columns where each row has some number of
labels and the overall cardinality of the labels is relatively low (the
index in-memory size will depend on the cardinality of the labels).

This PR also introduces the idea that different scalar indices may
support different types of queries. There is now `SargableQuery` (what
is used by btree and bitmap) and `LabelListQuery` (used by the new label
list index). Going forward we should probably break the FullTextQuery
out of `SargableQuery` and into its own query type.