perf: change scalar index to return RowIdTreeMap instead of u64 array (#2587)
The filtering paths that use scalar indices are all built to use
`RowIdTreeMap`. Scalar indices were built to use `UInt64Array`. This
means we have to copy from one to the other (an O(N log(N)) operation.
When we just had btree indices this was a fairly cheap operation since
the indices shouldn't be returning large #'s of rows anyways. Now that
we have `BITMAP` and `LABEL_LIST` we can have indices that return very
large #'s of rows and this copy was expensive. By changing the scalar
indices to use `RowIdTreeMap` directly we skip the conversion.

NOTE: any `LABEL_LIST` or `BITMAP` index that was previously created
will be invalid and will need to be recreated. Since we haven't had an
official release of these index types I am not considering this a
breaking change.