refactor: remove custom ProjectionExec (#2608)
We implemented our own custom projection `ExecutionPlan` that took a
schema to project. With the latest DataFusion functions (particularly
`named_struct`) we can now express this in terms of the built-in
`ProjectionExec`. This reduces the maintenance burden on us and ensures
the equivalents properties and optimizations rules from DataFusion work
well without much effort on our part.