feat: add version tags (#2482)
Closes #588

This PR takes a different course from what was proposed in #588 and the
subsequent PR #605. Rather than using the existing `tag` field in the
manifest, this PR adds a new directory at the root of the dataset called
`_refs/tags`. This new directory contains files, where each filename
maps to a JSON that contains the version number. When a user wants to
checkout a specific tag, we do a lookup to find the version number, and
then simply call `checkout_version` under-the-hood. By considering tags
outside of the manifests file, we get the following benefits (all of
which are consistent with the more familiar git tag):

- Allow users to freely create and delete tags retroactively.
- Allow multiple tags to point to the same underlying version.
- Easily enforce that tag names are unique.
- Allow users to re-use a previously deleted tags(so long as the name is
unique among all other tags at the time of reuse).

---------

Co-authored-by: Will Jones <willjones127@gmail.com>