feat: add backpressure to v2 I/O scheduler (#2683)
Adds backpressure to the I/O scheduler. The general problem is rather
tricky. If your I/O threads all pause because your I/O buffer is full
and then your decode threads are waiting for queued I/O tasks then you
will have a deadlock.

I/O priority should allow us to avoid this situation but it is something
that will need to be monitored, especially if users want to set really
small I/O buffer sizes. For this reason I haven't made any of the new
settings configurable (except for deadlock detection which we may turn
on for debugging purposes if people seem to have encountered a
deadlock).

One way you can hit this deadlock is to create a file with 10 pages
where each page has 10GB of data. Even a single read will fill up the
I/O buffer. We submit the reads in priority order but we have 8 I/O
threads and so they race to grab the permits.

As a result I've added a much needed option to split primitive arrays
into multiple pages if we are given huge chunks of data. The splitting
algorithm is not perfect and could also use some work (perhaps we can do
a sort of "binary splitting" where we continuously split the largest
chunk in half until all chunks are below a given size)

At the moment I think things are "safe enough" that this PR prevents
more problems (avoids OOM) than it introduces (deadlock in esoteric
cases).