feat: make default batch size configurable via environment variable (#2693)
There's a number of places (updates, compaction, etc.) where the default
batch size is used and it is not configurable. However, the default is
too large in many cases. We should probably eventually make all of these
operations configurable (batch size, readahead, etc.) or perhaps change
the default batch size to be based on data size but in the meantime this
provides a potential workaround.