chore: relax the check in test_index_cache_size (#2723)
It sometimes fails with:

```
 FAILED python/tests/test_vector_index.py::test_index_cache_size - assert 0.9889441728591919 > 0.99
 +  where 0.9889441728591919 = <built-in method index_cache_hit_rate of _lib._Dataset object at 0xfffe31904af0>()
 +    where <built-in method index_cache_hit_rate of _lib._Dataset object at 0xfffe31904af0> = <_lib._Dataset object at 0xfffe31904af0>.index_cache_hit_rate
 +      where <_lib._Dataset object at 0xfffe31904af0> = <lance.dataset.LanceDataset object at 0xfffe31930340>._ds
```

Since the failures are always on ARM I think the root cause might be the
relaxed nature of the counter. We could just make the counter and
acquire / consume counter but I don't think super accurate index stats
are all that important (indeed, metric stats like this are often the
example given for relaxed memory model).