perf: coalesce ids before executing take (#2680)
Late materialization is a great benefit when executing a highly
selective filter. However, if a filter is highly selective it means that
each input batch will probably only have a few matching rows. The
current implementation executes take for each filtered batch. E.g.
instead of a single call of `take(500, 10000, 300000)` we get three
calls `take(500)`, `take(10000)`, and `take(300000)`. This means:

 * We can't coalesce
 * More CPU overhead (many calls to take_ranges)
 * Very small output batches (user's batch size is not respected)

On cloud storage I see a 10x plus benefit in scan performance.

We have a benchmark for this (EDA search plot 4) which should assist
with preventing regression in the future:
https://bencher.dev/console/projects/weston-lancedb/plots