fix: don't panic when v2 scans end early (reader is dropped) (#2690)
If the reader is dropped during a scan then the scheduler thread is
ended after the decoder thread and the io thread. This means the
scheduler thread might get an error sending to a closed receiver.
Previously we would panic on these errors. With this change we simply
log a debug message and end early.