fix: fix v2 error that can happen when writing list<struct<...>> with many empty lists (#2762)
If the list offsets are bigger than the struct items (easy if there are
many lists) and there are multiple pages of offsets then the list
encoding needs to force a page from the items page. This causes `flush`
to be called multiple times which the struct encoder wasn't expecting.
This would lead the struct header column to have multiple pages which,
while not technically a problem, causes an assertion to fail on read.

We could relax the assertion but it is simpler to just not write that
kind of file in the first place.