feat: constant-time manifest lookup on object stores (#2798)
This introduces `ManifestNamingScheme` with a `V1` and `V2` variant.
`V1` is the existing naming scheme. `V2` uses a scheme optimized for
object storage listing mechanisms, making looking up the latest manifest
constant time.

**On S3, this makes `lance.dataset()` take only 125 ms, regardless of
how many versions of the dataset existed. Previously this time grew
linearly with number of versions.**

We also provide a method `migrate_manifest_paths_v2`. Because this
method alters the manifest path scheme, and agreement on the scheme is
critical for the write path, it's important is it not run while any read
or write operations happen on that table. That's why this migration
can't happen in the background like some other migrations.

Closes #2790