fix: ensure the v2 file writer does not mix up the order of a column (#2836)
Sometimes the v2 writer will split a batch into multiple pages. When it
does this it encodes those pages in parallel. It is possible those
encoding tasks finish out of order and the writer was then writing the
pages out of order. This mean the order of one column could get out of
sync with the order of another column.

---------

Co-authored-by: broccoliSpicy <93440049+broccoliSpicy@users.noreply.github.com>