chore: fix aws-integ-test from running out of disk space (#2851)
The AWS integration tests were failing because we run out of disk space.
We have about 30GB of disk space per job. Docker (setting up minio), the
rust cache, building the wheels, and running the tests all utilize quite
a bit of disk space.

As a (possibly short-term) solution we change the integration tests job
to run only the integration tests and not all of the tests.