feat: add a separate retry loop on object contents download (#2874)
`object_store` has automatic retries. However, these retries only apply
on the initial HTTP request (the sending and receiving of the headers.
The download of the response body is not retried in the same way.

This download was previously covered by our outer retry loop but this
had two problems:
* It was not configurable
* It was not obvious what the error was (the error would say "error
decoding response body".

The download retry count is now configurable and the error message has
been enhanced.