fix: limit/offset pushdown was not working correctly in the presence of deletions (#2895)
We need to push back the scan start and potentially grab more rows than
requested to make sure we fulfill the request. For v2 files we could
also convert the row range into multiple ranges instead of grabbing more
rows than required. This might be something to investigate in the
future.