feat: log the execution plan in a more compact fashion (#2906)
Currently we log the debug representation of the plan (twice!) every
time a plan is executed. This is way too much information (e.g. the
object store's debug information is repeated many times). In one
egregious case I was dealing with that had 8k fragments this was
generating 35MB of log data!

This PR changes to an indented display representation of the plan.