fix: use retry loop when loading deletion files (#2899)
When loading deletion files we were directly accessing
`ObjectStore::inner` which bypasses the retry loop. This makes loading
deletion files somewhat brittle and was causing issues in scans (which
can load a lot of deletion files at the same time they are loading lots
of other ranges).

This PR changes the deletion file load to use the `ObjectStore` and adds
convenience methods to make this easier in the future.

In the future we should also move this load to use a `FileScheduler`
instead of `ObjectStore` so that these deletion file loads do not
violate our max-iops-per-scan but that can be done later.