fix: v2 deadlock when there are multiple data files (#2923)
We read batches from data files in a round-robin fashion. With the old
priority mechanism all data in file 1 would be scheduled before file 2.
This means we'd deadlock if file 1 was bigger than the I/O buffer size.

Now, each fragment has a unique priority, but the priority within the
fragment is the same for all data files.

This should avoid deadlock as long as the I/O buffer size is larger than
any given batch (may want to issue warning or fail if we can detect when
the batch size is too large).