fix: `refine_factor` should be used in `search_in_partition` for IVF V2 Indices (#2933)
In testing we saw that `refine_factor` was not increasing recall even
when set to a large enough value that it should be doing brute force
search in V3 indices. After tracing through the code we found that
`refine_factor` was not being utilized in the individual partitions
`search_in_partition` function, so they could only return at most `k`
values to be refined. This change makes it so that the
`search_in_partition` function will return `refine_factor * k` values
for further refinement.