feat: allow scalar indices to be disabled for a query (#2937)
There are some cases today where scalar indices can make query
performance worse. This is usually when a scalar index search has many
results (e.g. "x != 7"). Ideally, we will fix our planner to not use the
scalar index in this case. However, even if we do that, there will
probably be cases where it would be nice (for debugging purposes if
nothing else) to be able to disable the scalar index for a query.