feat: reconnect pushdown to v2 (#2913)
There is still a bit more testing work to do before pushdown is fully
supported in v2 and until we start using `LanceDfFieldDecoderStrategy`
in the file reader it won't be accessible to users. However, this PR has
a number of structural refactors for v2 and is big enough as it is.

This adds a cache to the v2 schedulers. This is needed in this PR
because we want to use the cache to store zone maps. However, it will be
needed in future 2.1 work as well because we want to cache things like
"rows per chunk" and "dictionaries".

This adds an initialization routine to v2 schedulers. Again, this is
needed for zone maps but will also be used by 2.1 features.

Lastly, this PR does, in fact, reconnect the zone maps feature,
restoring blocks that had been commented out.