feat: add a blob encoding for large binary values (#2868)
Compared to the regular large binary encoding this will have
considerably less metadata (since there will be fewer pages). In the
future it will also be possible to read just the descriptions (allowing
for different ways of reading the data such as file objects).

The scheduling algorithm is also more conservative as well. We don't
fetch descriptions until we need them and we only read in one batch of
blob data at a time (instead of an entire page). This means the reader
will probably be slightly less performant (though, when reading values
this large, it shouldn't be too noticeable) but will use considerably
less RAM.