feat: cuda acceleration for PQ builds/assignments (#2946)
Currently if an accelerator is used, it's only used for IVF training and
assignments. This PR extends it to also run on PQ training &
assignments.
I benchmarked on a gcloud n1-standard-16 instance with an attached
nvidia T4, using the wikipedia dataset with 50 in-sample queries (so qps
will be a bit noisy).

Before:

![results_static_20240927_210452_plot_dataset_wikipedia_k_10](https://github.com/user-attachments/assets/1eabc795-e025-4bbf-b3cf-2d98e6057223)


After:

![results_static_20240927_202649_plot_dataset_wikipedia_k_10](https://github.com/user-attachments/assets/74887ca6-4fca-4c6b-b752-aab1c8128917)

There's some noise due to randomness, but these plots can be considered
to be essentially the same, except for the faster build time.


Update: I've verified that there are no regressions from the latest
changes.