feat: make it possible to add columns from existing data (#3010)
The existing `add_columns` capability works well when you have a UDF or
SQL function that you want Lance to run for you. However, there are
times when the new column values have already been determined (e.g. when
a cluster farmed out the calculation of the values). In this case we
want a simple way to add new columns from existing values.

---------

Co-authored-by: BubbleCal <bubble-cal@outlook.com>