fix: verify num_sub_vectors is valid before creating index (#3056)
This was being checked deep in the Rust code on the CPU path. However,
on the accelerator path it was creating the index which would then give
obscure errors when attempts were made to search it.

This PR adds a check early in the python code and adds unit tests for
the various ways indices can be created.