chore: cleanup decoder benches (#3118)
This ensures that the setup routines happen inside of the benchmark
function. Otherwise datagen will be run for all benchmarks during the
collection phase and the system will run out of RAM.

This also changes the routines to use `FileMetadataCache::no_cache` as
it turns out `moka` doesn't immediately free its cache space on drop
and, when used in a tight benchmark loop, this leads to running out of
memory.