fix: fix performance regression introduced during reader refactor (#3170)
In https://github.com/lancedb/lance/pull/2913 we added an initialization
routine to the scheduler. Unfortunately, this caused a problem. It put
`schedule_ranges` behind an await which means that we wouldn't schedule
ranges for all files in a scan immediately. Instead, we wouldn't begin
scheduling file X until we had reached nearly the last batch of file
X-1. This introduced stuttering in the read path and had a significant
effect on performance.

The fix here moves the initialization into the dedicated scheduler
thread and keeps the creation of the scheduler a synchronous task which
is actually a bit simpler anyways I think.