feat: enhance repdef utilities to handle empty / null lists (#3200)
Empty & null lists are interesting. If you have them then your final
repetition & definition buffers will have more items than you have in
your flattened array. This fact required a considerably reworking in how
we build and unravel rep/def buffers.

When building we record the position of the specials and then, when we
serialize into rep/def buffers, we insert these special values. When
unraveling we need to deal with the fact that certain rep/def values are
"invisible" to the current context in which we are unraveling.

In addition, we now need to start keeping track of the structure of each
layer of repetition in the page metadata. This helps us understand the
meaning behind different definition levels later when we are unraveling.

This PR adds the changes to the rep/def utilities. We still aren't
actually using repetition levels at all yet. That will come in future
PRs.