fix!: correctly handle nulls in btree and bitmap indices (#3211)
There were a few issues with our null handling in scalar indices.

First, it appears I assumed earlier that `X < NULL` and `X > NULL` would
always be false. However, in `arrow-rs` the ordering considers `NULL` to
be "the smallest value" and so `X < NULL` always evaluated to true. This
required some changes to the logic in the btree and bitmap indices.

Second, the btree index was still using the v1 file format because it
relied on the page size to keep track of the index's batch size. I've
instead made the batch size a configurable property (configurable in
code, not configurable by users) and made it so that btree can use the
v2 file format.

Finally, related to the above, I changed it so we now write v2 files for
all scalar indices, even if the dataset is a v1 dataset. I think that's
a reasonable decision at this point.

The logic to fallback and read the old v1 files was already in place (I
believe @BubbleCal added it back when working on inverted index) but I
added a migration test just to be sure we weren't breaking our btree /
bitmap support.

Users with existing bitmap indices will get the new correct behavior
without any changes.
Users with existing btree indices will get some of the new correct
behavior but will need to retrain their indices to get all of the
correct behavior.

BREAKING CHANGE: Bitmap and btree indices will no longer be readable by
older versions of Lance. This is not a "backwards compatibility change"
(no APIs or code will stop working) but rather a "forwards compatibility
change" (you need to be careful in a multi-verison deployment or if you
roll back)