ci: update python/Cargo.log on version bump (#3207)
When we create the version bump commit it currently updates the lock
file `Cargo.lock` to point to the new versions. I suspect it is the
`cargo ws version --no-git-commit -y --exact --force 'lance*' ${{
inputs.part }}` command that does this. However, we have two lock files,
and `python/Cargo.lock` is not updated. This PR adds a step to the
version bump to also update `python/Cargo.lock`.