feat: add the repetition index to the miniblock write path (#3208)
The repetition index is what will give us random access support when we
have list data. At a high level it stores the number of top-level rows
in each mini-block chunk. We can use this later to figure out which
chunks we need to read.

In reality things are a little more complicated because we don't mandate
that each chunk starts with a brand new row (e.g. a row can span
multiple mini-block chunks). This is useful because we eventually want
to support arbitrarily deep nested access. If we create not-so-mini
blocks in the presence of large lists then we introduce read
amplification we'd like to avoid.