fix: correctly copy null buffer when making deep copy (#3238)
In some situations an array could be sliced in such a way that the array
had no offset, but the array's null buffer did have an offset. In these
cases we were not deep copying the array correctly and the offset of the
null buffer was lost. This does mean, in some cases, the 2.0 writer
could write incorrect nulls. However, the input conditions would mean
that the user's data would have to originate from rust in such a way
that it was sliced like this. It would be impossible for batches from
the C data interface or from python to look like this.