refactor(python)!: simplify marshalling of `Fragment`, `DataFile`, `Operation`, `Transaction` (#3240)
BREAKING CHANGE: `DataFile.deletion_file` is now a property, not a
method.

For `Fragment` and `Operation`, we had a sort of intermediate `inner`
layer to handle translating between Rust struct and Python objects. This
worked fine in isolation, but once you need to convert at the top of a
hierarchy it became tedious. This was the case for `Transaction`.
`Transaction` contained an `Operation`, which could contain many
`Fragment`s, which contains many `DataFile`s.

These structures are primarily data holders, so they've been made into
`dataclasses`. A newtype wrapper struct `PyLance<T>` is used to provide
implementations of `FromPyObject` and `ToPyObject`. This makes
signatures more readable, and makes the wrappers thinner. For example,
instead of a special Python `FragmentMetadata` struct, we just have a
`PyLance<Fragment>`, where `Fragment` is from the Rust crate.